Migrate Card story to typescript (#756)
* 🚚 Change extension to tsx

* 📝 Adding props table for Card and sub components

Had to use the JSX.IntrinsicElements trick to make the arg tables show

* 📝 Add the rest of the Card stories as examples

* 🚚 Rename function without eds prefic

* 🔥 Remove knobs completely

* ➕ Re-add types for storybook, react and react-dom
