+++ libraries/core-react/src/Card/CardHeaderTitle.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef } from 'react'
 import styled from 'styled-components'
 
-type Props = React.HTMLAttributes<HTMLDivElement>
+export type CardHeaderTitleProps = JSX.IntrinsicElements['div']
 
 const StyledCardHeaderTitle = styled.div`
   display: grid;
@@ -9,8 +9,8 @@ const StyledCardHeaderTitle = styled.div`
   grid-auto-columns: auto;
 `
 
-export const CardHeaderTitle = forwardRef<HTMLDivElement, Props>(
-  function EdsCardHeaderTitle({ children, className = '', ...rest }, ref) {
+export const CardHeaderTitle = forwardRef<HTMLDivElement, CardHeaderTitleProps>(
+  function CardHeaderTitle({ children, className = '', ...rest }, ref) {
     const props = {
       ...rest,
       className,
