+++ libraries/core-react/src/Card/index.ts
@@ -4,18 +4,18 @@ import { CardMedia } from './CardMedia'
 import { CardHeader } from './CardHeader'
 import { CardHeaderTitle } from './CardHeaderTitle'
 
-type CardType = typeof CardWrapper & {
+type CardProps = typeof CardWrapper & {
   CardActions: typeof CardActions
   CardHeader: typeof CardHeader
   CardMedia: typeof CardMedia
   CardHeaderTitle: typeof CardHeaderTitle
 }
 
-const Card = CardWrapper as CardType
+const Card = CardWrapper as CardProps
 
 Card.CardActions = CardActions
 Card.CardHeader = CardHeader
 Card.CardMedia = CardMedia
 Card.CardHeaderTitle = CardHeaderTitle
 
-export { Card }
+export { Card, CardProps }
