+++ libraries/core-react/src/index.ts
@@ -8,8 +8,8 @@ export { Icon } from './Icon'
 export { List } from './List'
 export * from './Accordion'
 export { Tabs } from './Tabs'
+export * from './Card'
 export * from './TopBar'
-export { Card } from './Card'
 export { Dialog } from './Dialog'
 export { Scrim } from './Scrim'
 export { TableOfContents } from './TableOfContents'
